authorization added in all apis in charges

+++ src/routes/chargesRoutes.js
@@ -1,15 +1,31 @@
 import express from "express";
 import { chargesRegister } from "../models/ChargesRegisterModel.js";
 import { v4 as uuidv4 } from 'uuid';
+import { authenticate } from "../middleware/middleware.js";
+import { getUserIdFromToken } from "../helper/generateToken.js";
 
 const router = express.Router();
 
-router.post("/api/usercharge/addChargeRegister", async (req, res) => {
+router.post("/api/usercharge/addChargeRegister", authenticate, async (req, res) => {
     try {
+        const token = req.header('Authorization');
+        const userId = getUserIdFromToken(token)
         const chargeRegData = req.body;
         const chargeId = uuidv4();
         chargeRegData["chargeRegId"] = chargeId;
-        const createChargeData = await chargesRegister.create(chargeRegData);
+        const { typesOfCharges, chargesHeading, printAs, accountHeadToPost, typesOfCharges1, inputAmountOfChargesAs, taxSettings } = req.body;
+        const craeteChargesData = {
+            userId: userId,
+            chargeRegId : chargeId,
+            chargesHeading: chargesHeading,
+            printAs: printAs,
+            accountHeadToPost: accountHeadToPost,
+            typesOfCharges: typesOfCharges,
+            typesOfCharges1: inputAmountOfChargesAs === "percentage" ? typesOfCharges1 : null,
+            inputAmountOfChargesAs: inputAmountOfChargesAs,
+            taxSettings: taxSettings
+        };
+        const createChargeData = await chargesRegister.create(craeteChargesData);
         if (createChargeData) {
             return res.status(200).json({ success: true, message: "Charge Registered Successfully!", chargeRegId: createChargeData.chargeRegId });
         }
@@ -18,9 +34,11 @@ router.post("/api/usercharge/addChargeRegister", async (req, res) => {
     }
 });
 
-router.post("/api/usercharge/getAllChargeList", async (req, res) => {
+router.post("/api/usercharge/getAllChargeList", authenticate, async (req, res) => {
     try {
-        const findAllChargeList = await chargesRegister.find({});
+        const token = req.header('Authorization');
+        const userId = getUserIdFromToken(token)
+        const findAllChargeList = await chargesRegister.find({ userId: userId },{ __v:0, _id:0, createdAt:0, updatedAt:0 });
         if (findAllChargeList) {
             return res.status(200).json({ success: true, message: findAllChargeList });
         } else {
@@ -31,11 +49,13 @@ router.post("/api/usercharge/getAllChargeList", async (req, res) => {
     }
 });
 
-router.post("/api/usercharge/getChargeData", async (req, res) => {
+router.post("/api/usercharge/getChargeData", authenticate, async (req, res) => {
     try {
+        const token = req.header('Authorization');
+        const userId = getUserIdFromToken(token)
         const chargeRegData = req.body;
         const getChargeId = chargeRegData.chargeRegId;
-        const findChargeId = await chargesRegister.findOne({ chargeRegId: getChargeId });
+        const findChargeId = await chargesRegister.findOne({ chargeRegId: getChargeId, userId: userId },{ __v:0, _id:0, createdAt:0, updatedAt:0 });
         if (findChargeId) {
             return res.status(200).json({ success: true, message: findChargeId });
         } else {
@@ -46,11 +66,14 @@ router.post("/api/usercharge/getChargeData", async (req, res) => {
     }
 });
 
-router.put("/api/usercharge/updateChargeData", async (req, res) => {
+router.put("/api/usercharge/updateChargeData", authenticate, async (req, res) => {
     try {
+        const token = req.header('Authorization');
+        const userId = getUserIdFromToken(token)
         const chargeRegData = req.body;
-        const getChargeId = chargeRegData.chargeRegId;
-        const updateChargeId = await chargesRegister.findOneAndUpdate({ chargeRegId: getChargeId }, { $set: chargeRegData });
+        const { typesOfCharges1, chargeRegId } = req.body;
+        chargeRegData["typesOfCharges1"] = chargeRegData.inputAmountOfChargesAs === "percentage" ? typesOfCharges1 : null;
+        const updateChargeId = await chargesRegister.findOneAndUpdate({ chargeRegId: chargeRegId, userId: userId }, { $set: chargeRegData });
         if (updateChargeId) {
             return res.status(200).json({ success: true, message: "Charge Register Id Updated Successfully!" });
         } else {
@@ -61,11 +84,13 @@ router.put("/api/usercharge/updateChargeData", async (req, res) => {
     }
 });
 
-router.delete("/api/usercharge/deleteChargeData", async (req, res) => {
+router.delete("/api/usercharge/deleteChargeData", authenticate, async (req, res) => {
     try {
+        const token = req.header('Authorization');
+        const userId = getUserIdFromToken(token)
         const chargeRegData = req.body;
         const getChargeId = chargeRegData.chargeRegId;
-        const findChargeId = await chargesRegister.findOneAndDelete({ chargeRegId: getChargeId });
+        const findChargeId = await chargesRegister.findOneAndDelete({ chargeRegId: getChargeId, userId: userId });
         if (findChargeId) {
             return res.status(200).json({ success: true, message: "Charge Register Id Deleted Successfully!" });
         } else {
